use PostgreSQL's own hash_any function as default partfunc; remove all murmur3-related source code. Alter regression tests to reflect new hash values for inputs, and a slightly different set of input data to ensure that sufficient chunks and partitions are tested. Some changes to .sh scripts in sql/setup that seem to be used only to power the "unit tests", which I cannot yet run successfully.
