+++ sql/setup/set_meta.sh
@@ -16,12 +16,11 @@ if [ "$#" -ne 2 ] ; then
     exit 1
 fi
 
-NODENAME=$1
 NODEHOST=$2
 
 echo "Connecting to $POSTGRES_HOST as user $POSTGRES_USER and with db $INSTALL_DB"
 psql -U $POSTGRES_USER -h $POSTGRES_HOST -d $INSTALL_DB -v ON_ERROR_STOP=1  <<EOF
-SELECT set_meta('$NODENAME' :: NAME, '$NODEHOST'::text);
+SELECT set_meta('$NODEHOST'::text);
 EOF
 
 cd $PWD
