+++ test/expected/copy_from.out
@@ -31,19 +31,21 @@ COMMIT;
 INSERT INTO public."two_Partitions"("timeCustom", device_id, series_0, series_1) VALUES
 (1257987600000000000, 'dev1', 1.5, 1),
 (1257987600000000000, 'dev1', 1.5, 2),
-(1257894000000000000, 'dev20', 1.5, 1),
+(1257894000000000000, 'dev2', 1.5, 1),
 (1257894002000000000, 'dev1', 2.5, 3);
 INSERT INTO "two_Partitions"("timeCustom", device_id, series_0, series_1) VALUES
-(1257894000000000000, 'dev20', 1.5, 2);
+(1257894000000000000, 'dev2', 1.5, 2);
 \o
-COPY (SELECT * FROM "two_Partitions" ) TO STDOUT;
+COPY (SELECT * FROM "two_Partitions" ORDER BY "timeCustom", device_id) TO STDOUT;
 1257894000000000000	dev1	1.5	1	2	t
 1257894000000000000	dev1	1.5	2	\N	\N
+1257894000000000000	dev2	1.5	2	\N	\N
+1257894000000000000	dev2	1.5	1	\N	\N
 1257894000000001000	dev1	2.5	3	\N	\N
 1257894001000000000	dev1	3.5	4	\N	\N
 1257894002000000000	dev1	2.5	3	\N	\N
+1257894002000000000	dev1	5.5	6	\N	t
+1257894002000000000	dev1	5.5	7	\N	f
 1257897600000000000	dev1	4.5	5	\N	f
-1257987600000000000	dev1	1.5	1	\N	\N
 1257987600000000000	dev1	1.5	2	\N	\N
-1257894000000000000	dev20	1.5	1	\N	\N
-1257894000000000000	dev20	1.5	2	\N	\N
+1257987600000000000	dev1	1.5	1	\N	\N
