+++ test/expected/create_chunks.out
@@ -41,23 +41,24 @@ SELECT * FROM set_chunk_time_interval('chunk_test', 40::bigint);
 (1 row)
 
 INSERT INTO chunk_test VALUES(23, 3, 'dev3');
-SELECT * FROM chunk_test;
+SELECT * FROM chunk_test order by time, metric, device_id;
  time | metric | device_id 
 ------+--------+-----------
     1 |      1 | dev1
     2 |      2 | dev2
+   23 |      3 | dev3
    45 |      2 | dev2
    46 |      2 | dev2
-   23 |      3 | dev3
 (5 rows)
 
 SELECT * FROM _timescaledb_catalog.chunk;
  id | partition_id | start_time | end_time 
 ----+--------------+------------+----------
   1 |            1 |          0 |        9
-  2 |            1 |         40 |       49
-  3 |            1 |         10 |       39
-(3 rows)
+  2 |            2 |          0 |        9
+  3 |            2 |         40 |       49
+  4 |            1 |         10 |       39
+(4 rows)
 
 SELECT * FROM _timescaledb_catalog.hypertable;
  id | schema_name | table_name | associated_schema_name | associated_table_prefix |   root_schema_name    | root_table_name | replication_factor | placement | time_column_name | time_column_type | created_on | chunk_time_interval 
@@ -74,11 +75,13 @@ SELECT * FROM _timescaledb_catalog.chunk c
     LEFT JOIN _timescaledb_catalog.chunk_replica_node crn ON (c.id = crn.chunk_id)
     LEFT JOIN _timescaledb_catalog.partition_replica pr ON (crn.partition_replica_id = pr.id)
     LEFT JOIN _timescaledb_catalog.hypertable h ON (pr.hypertable_id = h.id)
-    WHERE h.schema_name = 'public' AND h.table_name = 'chunk_test';
+    WHERE h.schema_name = 'public' AND h.table_name = 'chunk_test'
+    ORDER BY c.id;
  id | partition_id | start_time | end_time | chunk_id | partition_replica_id | database_name |      schema_name      |     table_name      | id | partition_id | hypertable_id | replica_id |      schema_name      |       table_name       | id | schema_name | table_name | associated_schema_name | associated_table_prefix |   root_schema_name    | root_table_name | replication_factor | placement | time_column_name | time_column_type | created_on | chunk_time_interval 
 ----+--------------+------------+----------+----------+----------------------+---------------+-----------------------+---------------------+----+--------------+---------------+------------+-----------------------+------------------------+----+-------------+------------+------------------------+-------------------------+-----------------------+-----------------+--------------------+-----------+------------------+------------------+------------+---------------------
   1 |            1 |          0 |        9 |        1 |                    1 | single        | _timescaledb_internal | _hyper_1_1_0_1_data |  1 |            1 |             1 |          0 | _timescaledb_internal | _hyper_1_1_0_partition |  1 | public      | chunk_test | _timescaledb_internal  | _hyper_1                | _timescaledb_internal | _hyper_1_root   |                  1 | STICKY    | time             | bigint           | single     |                  40
-  2 |            1 |         40 |       49 |        2 |                    1 | single        | _timescaledb_internal | _hyper_1_1_0_2_data |  1 |            1 |             1 |          0 | _timescaledb_internal | _hyper_1_1_0_partition |  1 | public      | chunk_test | _timescaledb_internal  | _hyper_1                | _timescaledb_internal | _hyper_1_root   |                  1 | STICKY    | time             | bigint           | single     |                  40
-  3 |            1 |         10 |       39 |        3 |                    1 | single        | _timescaledb_internal | _hyper_1_1_0_3_data |  1 |            1 |             1 |          0 | _timescaledb_internal | _hyper_1_1_0_partition |  1 | public      | chunk_test | _timescaledb_internal  | _hyper_1                | _timescaledb_internal | _hyper_1_root   |                  1 | STICKY    | time             | bigint           | single     |                  40
-(3 rows)
+  2 |            2 |          0 |        9 |        2 |                    2 | single        | _timescaledb_internal | _hyper_1_2_0_2_data |  2 |            2 |             1 |          0 | _timescaledb_internal | _hyper_1_2_0_partition |  1 | public      | chunk_test | _timescaledb_internal  | _hyper_1                | _timescaledb_internal | _hyper_1_root   |                  1 | STICKY    | time             | bigint           | single     |                  40
+  3 |            2 |         40 |       49 |        3 |                    2 | single        | _timescaledb_internal | _hyper_1_2_0_3_data |  2 |            2 |             1 |          0 | _timescaledb_internal | _hyper_1_2_0_partition |  1 | public      | chunk_test | _timescaledb_internal  | _hyper_1                | _timescaledb_internal | _hyper_1_root   |                  1 | STICKY    | time             | bigint           | single     |                  40
+  4 |            1 |         10 |       39 |        4 |                    1 | single        | _timescaledb_internal | _hyper_1_1_0_4_data |  1 |            1 |             1 |          0 | _timescaledb_internal | _hyper_1_1_0_partition |  1 | public      | chunk_test | _timescaledb_internal  | _hyper_1                | _timescaledb_internal | _hyper_1_root   |                  1 | STICKY    | time             | bigint           | single     |                  40
+(4 rows)
