+++ test/expected/insert_single.out
@@ -39,10 +39,10 @@ COMMIT;
 INSERT INTO "one_Partition"("timeCustom", device_id, series_0, series_1) VALUES
 (1257987600000000000, 'dev1', 1.5, 1),
 (1257987600000000000, 'dev1', 1.5, 2),
-(1257894000000000000, 'dev20', 1.5, 1),
+(1257894000000000000, 'dev2', 1.5, 1),
 (1257894002000000000, 'dev1', 2.5, 3);
 INSERT INTO "one_Partition"("timeCustom", device_id, series_0, series_1) VALUES
-(1257894000000000000, 'dev20', 1.5, 2);
+(1257894000000000000, 'dev2', 1.5, 2);
 \d+ "one_Partition".*
 Index "one_Partition.1-one_Partition_device_id_timeCustom_idx"
    Column   |  Type  |  Definition  | Storage  
@@ -243,20 +243,22 @@ Child tables: "one_Partition"._hyper_1_1_0_1_data,
  series_bool | boolean          |           | plain    |              | 
 Child tables: "one_Partition"._hyper_1_0_replica
 
-SELECT * FROM "one_Partition";
+SELECT * FROM "one_Partition" ORDER BY "timeCustom", device_id;
      timeCustom      | device_id | series_0 | series_1 | series_2 | series_bool 
 ---------------------+-----------+----------+----------+----------+-------------
  1257894000000000000 | dev1      |      1.5 |        1 |        2 | t
  1257894000000000000 | dev1      |      1.5 |        2 |          | 
+ 1257894000000000000 | dev2      |      1.5 |        1 |          | 
+ 1257894000000000000 | dev2      |      1.5 |        2 |          | 
  1257894000000001000 | dev1      |      2.5 |        3 |          | 
  1257894001000000000 | dev1      |      3.5 |        4 |          | 
- 1257894000000000000 | dev20     |      1.5 |        1 |          | 
  1257894002000000000 | dev1      |      2.5 |        3 |          | 
- 1257894000000000000 | dev20     |      1.5 |        2 |          | 
+ 1257894002000000000 | dev1      |      5.5 |        6 |          | t
+ 1257894002000000000 | dev1      |      5.5 |        7 |          | f
  1257897600000000000 | dev1      |      4.5 |        5 |          | f
  1257987600000000000 | dev1      |      1.5 |        1 |          | 
  1257987600000000000 | dev1      |      1.5 |        2 |          | 
-(10 rows)
+(12 rows)
 
 --test that we can insert data into a 1-dimensional table (only time partitioning)
 CREATE TABLE "1dim"(time timestamp, temp float);
