+++ test/sql/copy_from.sql
@@ -2,6 +2,6 @@
 \ir include/insert_two_partitions.sql
 \o
 
-COPY (SELECT * FROM "two_Partitions" ) TO STDOUT;
+COPY (SELECT * FROM "two_Partitions" ORDER BY "timeCustom", device_id) TO STDOUT;
