+++ test/sql/data/ds1_dev1_1.tsv
@@ -3,3 +3,5 @@
 1257894000000001000	dev1	2.5	3		
 1257894001000000000	dev1	3.5	4		
 1257897600000000000	dev1	4.5	5		false
+1257894002000000000	dev1	5.5	6		true
+1257894002000000000	dev1	5.5	7		false
