+++ test/sql/insert.sql
@@ -2,12 +2,12 @@
 
 \d+ "_timescaledb_internal".*
 SELECT *
-FROM "_timescaledb_internal"._hyper_1_0_replica;
+FROM "_timescaledb_internal"._hyper_1_0_replica order by "timeCustom", device_id;
 SELECT * FROM _timescaledb_catalog.chunk;
 SELECT * FROM _timescaledb_catalog.chunk_replica_node;
 SELECT * FROM _timescaledb_catalog.partition_replica;
 
-SELECT * FROM "two_Partitions";
+SELECT * FROM "two_Partitions" ORDER BY "timeCustom", device_id;
 SELECT * FROM ONLY "two_Partitions";
 
 CREATE TABLE error_test(time timestamp, temp float8, device text NOT NULL);
