+++ test/sql/update.sql
@@ -2,9 +2,9 @@
 \ir include/insert_single.sql
 \o
 
-SELECT * FROM "one_Partition";
+SELECT * FROM "one_Partition" ORDER BY "timeCustom", device_id;
 
 UPDATE "one_Partition" SET series_1 = 47;
 UPDATE "one_Partition" SET series_bool = true;
-SELECT * FROM "one_Partition";
+SELECT * FROM "one_Partition" ORDER BY "timeCustom", device_id;
